node-python-master/requirements.txt
grpcio==1.46.1
grpcio-tools==1.46.1
absl-py==1.0.0
